Split out manifest into separate file
